pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "optical-constant-plugin"
version = "0.1.0"
description = "Optical constant plugin"
authors = [{name = "Gemma"}]
requires-python = ">=3.10"
dependencies = ["nomad-lab"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points."nomad.plugin"]
optical_schema = "optical_constant_plugin.schema_packages.mypackage:optical_schema"
optical_parser = "optical_constant_plugin.parsers.myparsers:optical_parser"
optical_normalizer = "optical_constant_plugin.normalizers.mynormalizer:optical_normalizer"
optical_app        = "optical_constant_plugin.apps.myapp:optical_app"
electrical_parser = "optical_constant_plugin.parsers.myparsers:electrical_parser"
electrical_normalizer = "optical_constant_plugin.normalizers.mynormalizer:electrical_normalizer"
electrical_app = "optical_constant_plugin.apps.myapp:electrical_app"
materials_app = "optical_constant_plugin.apps.myapp:materials_app"

